abstracted the db init fn to utils folder
