+++ app/_layout.tsx
@@ -5,6 +5,7 @@ import { SQLiteProvider } from 'expo-sqlite';
 import { StatusBar } from 'expo-status-bar';
 import 'react-native-reanimated';
 
+import dbInit from '@/components/db-utils/InitDatabase';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 export default function RootLayout() {
@@ -17,22 +18,6 @@ export default function RootLayout() {
     // Async font loading only occurs in development.
     return null;
   }
-
-  // TODO: abstract this to utils/db folder
-const dbInit = async (db: any) => {
-  await db.execAsync(`
-    CREATE TABLE IF NOT EXISTS users (
-      id INTEGER PRIMARY KEY AUTOINCREMENT,
-      username TEXT NOT NULL,
-      password TEXT NOT NULL,
-      email TEXT NOT NULL UNIQUE,
-      first_name TEXT NOT NULL,
-      last_name TEXT NOT NULL
-    );
-  `);
-  
-  await db.execAsync(`PRAGMA journal_mode = WAL;`);
-}
 
   return (
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
